requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
python-dateutil==2.8.2
parsedatetime==2.6
requests==2.31.0
python-dotenv==1.0.0 